requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
numpy<2
matplotlib
Pillow
requests
paho-mqtt
sense-hat
picamera2
